requirements.txt
nose==1.3.0
PyHamcrest==1.7.2
funcy==0.5.6
